move Fight declaration out of PhaseParser
